@AngelicosPhosphoros 2021-12-20
@the8472 @nrc

Rewritten benchmarks and split my benchmarks into 2 parks.
My optimization mostly implemented to improve `vec::bench_dedup_none`.

Here the results. Affected benchmarks is `vec::bench_dedup_all`, `vec::bench_dedup_none`, `vec::bench_dedup_random`. `vec::bench_dedup_slice_truncate` shown to show detect possible measuring errors.

### Old code benchmark
```
test vec::bench_dedup_all_100                ... bench:          57 ns/iter (+/- 2) = 7017 MB/s
test vec::bench_dedup_all_1000               ... bench:         394 ns/iter (+/- 5) = 10152 MB/s
test vec::bench_dedup_all_10000              ... bench:       4,019 ns/iter (+/- 6) = 9952 MB/s
test vec::bench_dedup_all_100000             ... bench:      39,567 ns/iter (+/- 183) = 10109 MB/s
test vec::bench_dedup_none_100               ... bench:          56 ns/iter (+/- 0) = 7142 MB/s
test vec::bench_dedup_none_1000              ... bench:         486 ns/iter (+/- 2) = 8230 MB/s
test vec::bench_dedup_none_10000             ... bench:       4,824 ns/iter (+/- 12) = 8291 MB/s
test vec::bench_dedup_none_100000            ... bench:      48,140 ns/iter (+/- 95) = 8309 MB/s
test vec::bench_dedup_random_100             ... bench:          64 ns/iter (+/- 3) = 6250 MB/s
test vec::bench_dedup_random_1000            ... bench:         779 ns/iter (+/- 11) = 5134 MB/s
test vec::bench_dedup_random_10000           ... bench:       9,968 ns/iter (+/- 43) = 4012 MB/s
test vec::bench_dedup_random_100000          ... bench:     365,905 ns/iter (+/- 1,113) = 1093 MB/s
test vec::bench_dedup_slice_truncate_100     ... bench:          91 ns/iter (+/- 4) = 4395 MB/s
test vec::bench_dedup_slice_truncate_1000    ... bench:         766 ns/iter (+/- 39) = 5221 MB/s
test vec::bench_dedup_slice_truncate_10000   ... bench:      14,793 ns/iter (+/- 109) = 2703 MB/s
test vec::bench_dedup_slice_truncate_100000  ... bench:     402,150 ns/iter (+/- 10,161) = 994 MB/s
```

### New code benchmark
```
test vec::bench_dedup_all_100                ... bench:          51 ns/iter (+/- 1) = 7843 MB/s
test vec::bench_dedup_all_1000               ... bench:         359 ns/iter (+/- 129) = 11142 MB/s
test vec::bench_dedup_all_10000              ... bench:       3,562 ns/iter (+/- 8) = 11229 MB/s
test vec::bench_dedup_all_100000             ... bench:      36,504 ns/iter (+/- 83) = 10957 MB/s
test vec::bench_dedup_none_100               ... bench:          44 ns/iter (+/- 0) = 9090 MB/s
test vec::bench_dedup_none_1000              ... bench:         288 ns/iter (+/- 1) = 13888 MB/s
test vec::bench_dedup_none_10000             ... bench:       2,752 ns/iter (+/- 28) = 14534 MB/s
test vec::bench_dedup_none_100000            ... bench:      29,269 ns/iter (+/- 1,082) = 13666 MB/s
test vec::bench_dedup_random_100             ... bench:          58 ns/iter (+/- 4) = 6896 MB/s
test vec::bench_dedup_random_1000            ... bench:         775 ns/iter (+/- 17) = 5161 MB/s
test vec::bench_dedup_random_10000           ... bench:      10,686 ns/iter (+/- 979) = 3743 MB/s
test vec::bench_dedup_random_100000          ... bench:     374,580 ns/iter (+/- 1,192) = 1067 MB/s
test vec::bench_dedup_slice_truncate_100     ... bench:          93 ns/iter (+/- 0) = 4301 MB/s
test vec::bench_dedup_slice_truncate_1000    ... bench:         970 ns/iter (+/- 5) = 4123 MB/s
test vec::bench_dedup_slice_truncate_10000   ... bench:      15,095 ns/iter (+/- 57) = 2649 MB/s
test vec::bench_dedup_slice_truncate_100000  ... bench:     389,795 ns/iter (+/- 293) = 1026 MB/s
```

It seems I managed to significantly improve `vec::bench_dedup_none_100000` without hurting others.